+++ ui/helpers/constants.js
@@ -253,6 +253,16 @@ export const phCompareDefaultOriginalRepo = 'mozilla-central';
 
 export const phCompareDefaultNewRepo = 'try';
 
+export const phTimeRanges = [
+  { value: 86400, text: 'Last day' },
+  { value: 86400 * 2, text: 'Last 2 days' },
+  { value: 604800, text: 'Last 7 days' },
+  { value: 1209600, text: 'Last 14 days' },
+  { value: 2592000, text: 'Last 30 days' },
+  { value: 5184000, text: 'Last 60 days' },
+  { value: 7776000, text: 'Last 90 days' },
+  { value: 31536000, text: 'Last year' }];
+
 export const phDefaultTimeRangeValue = 1209600;
 
 export const phTimeRangeValues = {
