+++ ui/js/controllers/perf/alerts.js
@@ -11,6 +11,7 @@ import editAlertSummaryNotesCtrlTemplate from '../../../partials/perf/editnotesc
 import { getApiUrl, getJobsUrl } from '../../../helpers/url';
 import {
   thDateFormat,
+  phTimeRanges,
   phDefaultTimeRangeValue,
   phTimeRangeValues,
   phAlertSummaryStatusMap,
@@ -19,7 +20,6 @@ import {
 import OptionCollectionModel from '../../../models/optionCollection';
 import PushModel from '../../../models/push';
 import RepositoryModel from '../../../models/repository';
-import { phTimeRanges } from '../../../perfherder/constants';
 
 perf.factory('PhBugs', [
     '$http', '$httpParamSerializer', '$interpolate', '$rootScope', 'dateFilter',
